I fixed the way we were adding edges. We need to look at all combinations, we cannot simply delete an edge betweeen the two sets if it induces a forbidden subgraph. This is because this edge may exist with other edges which have not been added yet!
Tonight, I will also redo the way we compare the clique number and
chromatic number. Instead of running two costly algorithms, we can
simply compute the clique number (which is hard), then construct a
k-coloring for Omega = k and see if this colors the graph in question.
This is a good approach since we conjecture the two numbers are equal
anyway.